no-std "port" of the unified logger
Basically I moved away the memory mapped file implementation.

We still need a proper implementation or traits to implement that on
embedded devices.
